pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "claude"
authors = [{name = "John Kitchin", email = "jkitchin@andrew.cmu.edu"}]
dependencies = [
	     "waitress",
	     "Adafruit-Blinka",
	     "adafruit-circuitpython-as7341",
	     "lgpio",
	     "gpiozero",
	     "flask",
	     "jsonlines",
	     "matplotlib",
	     "pandas",
	     "python-dotenv"]
license = {text = "MIT"}
dynamic = ["version"]

[project.scripts]
claude = "claude.app:run"